[spatialPartitioning] Move Knn traits as NeighborGraph Traits
