+++ Ponca/src/SpatialPartitioning/KnnGraph/knnGraph.h
@@ -6,15 +6,13 @@
 
 #pragma once
 
-#include "./knnGraphTraits.h"
+#include "../NeighborGraph/neighborGraph.h"
 
 #include "Query/knnGraphKNearestQuery.h"
 #include "Query/knnGraphRangeQuery.h"
 
 #include "../KdTree/kdTree.h"
 #include "../../Common/Assert.h"
-
-#include <memory>
 
 namespace Ponca
 {
@@ -27,95 +25,11 @@ namespace Ponca
      *
      * Provides default implementation of the KnnGraph
      *
-     * \see KnnGraphDefaultTraits for the default trait interface documentation.
+     * \see NeighborGraphDefaultTraits for the default trait interface documentation.
      * \see KnnGraphBase for complete API
      */
     template <typename DataPoint>
-    using KnnGraph = KnnGraphBase<KnnGraphDefaultTraits<DataPoint>>;
-
-    /// \brief Internal structure storing the buffers used by a neighbor graph
-    template <typename _Traits>
-    struct NeighborGraphBufferBase
-    {
-        using Traits         = _Traits;
-        using PointContainer = typename Traits::PointContainer;
-        using IndexContainer = typename Traits::IndexContainer;
-
-        PointContainer points;  ///< Buffer storing the input points (read only)
-        IndexContainer indices; ///< Buffer storing the indices associating the input points to the nodes
-
-        size_t points_size{0};
-        size_t indices_size{0};
-
-        PONCA_MULTIARCH inline NeighborGraphBufferBase() = default;
-        PONCA_MULTIARCH inline NeighborGraphBufferBase(PointContainer _points) : points(_points) {}
-
-        PONCA_MULTIARCH inline NeighborGraphBufferBase(PointContainer _points,
-                                                       typename Traits::IndexContainerRef _indices,
-                                                       const size_t _points_size, const size_t _indices_size)
-            : points(_points), indices(_indices), points_size(_points_size), indices_size(_indices_size)
-        {
-        }
-    };
-
-    /*! \brief Base class for neighbor classes
-     *
-     * \tparam _Traits Traits type providing the types and constants used by the KnnGraph. Must have the
-     * same interface as the default traits type.
-     * \tparam BufferType Type of buffer used in the Graph. Must inherit NeighborGraphBufferBase and be templated by
-     * _Traits
-     */
-    template <typename _Traits, template <typename> typename BufferType>
-    class NeighborGraphBase
-    {
-    public:
-#define WRITE_TRAITS                                                                                                    \
-    using Traits            = _Traits;                        /*!< Alias to the Traits type                         */  \
-    using DataPoint         = typename Traits::DataPoint;     /*!< DataPoint given by user via Traits               */  \
-    using Scalar            = typename DataPoint::Scalar;     /*!< Scalar given by user via DataPoint               */  \
-    using VectorType        = typename DataPoint::VectorType; /*!< VectorType given by user via DataPoint           */  \
-    using IndexType         = typename Traits::IndexType;     /*!< Type used to index points into the PointContainer*/  \
-    using PointContainer    = typename Traits::PointContainer; /*!< Container for DataPoint used inside the KdTree   */ \
-    using IndexContainer    = typename Traits::IndexContainer; /*!< Container for indices used inside the KdTree     */ \
-    using IndexContainerRef = typename Traits::IndexContainerRef; /*!< Ref type to index container                */
-        WRITE_TRAITS
-
-        using Buffers = BufferType<Traits>;
-        static_assert(std::is_base_of_v<NeighborGraphBufferBase<Traits>, Buffers>,
-                      "BufferType must inherit NeighborGraphBufferBase");
-
-    protected:
-        PONCA_MULTIARCH inline const IndexType* getIndexPtr() const { return Traits::getIndexRawPtr(m_bufs.indices); }
-        PONCA_MULTIARCH inline IndexType* getIndexPtr() { return Traits::getIndexRawPtr(m_bufs.indices); }
-
-    public:
-        // Data --------------------------------------------------------------------
-
-        /*! \brief Constructor that allows the use of prebuilt KnnGraph containers.
-         *
-         * Each internal values of a KnnGraph can be extracted using \ref `KnnGraph::buffers()`
-         *
-         * \note This constructor can be used to avoid the convertion and building process,
-         * which is useful to transfer directly the KnnGraph to the device in CUDA.
-         *
-         * \param _bufs Internal buffers of the KnnGraph
-         */
-        PONCA_MULTIARCH inline NeighborGraphBase(const Buffers& _bufs) : m_bufs(_bufs) {}
-
-        //! \brief Get the number of indices
-        PONCA_MULTIARCH [[nodiscard]] inline IndexType sampleCount() const { return (IndexType)m_bufs.indices_size; }
-        //! \brief Get the number of points
-        PONCA_MULTIARCH [[nodiscard]] inline IndexType pointCount() const { return (IndexType)m_bufs.points_size; }
-        //! \brief Get the internal point container
-        PONCA_MULTIARCH [[nodiscard]] inline PointContainer points() const { return m_bufs.points; };
-        //! \brief Get the internal index container
-        PONCA_MULTIARCH [[nodiscard]] inline IndexContainer samples() const { return m_bufs.indices; };
-        //! \brief Get access to the internal buffer, for instance to prepare GPU binding
-        PONCA_MULTIARCH [[nodiscard]] inline const Buffers& buffers() const { return m_bufs; }
-
-    protected:          // for friends relations
-        Buffers m_bufs; ///< Buffers used to store the KnnGraph
-    };
+    using KnnGraph = KnnGraphBase<NeighborGraphDefaultTraits<DataPoint>>;
 
     /// \brief Buffer class for StaticKnnGraphBase.
     ///
@@ -123,6 +37,15 @@ namespace Ponca
     template <typename _Traits>
     struct KnnGraphBuffers : public NeighborGraphBufferBase<_Traits>
     {
+#define WRITE_TRAITS                                                                                                    \
+    using Traits            = _Traits;                    /*!< Alias to the Traits type                          */     \
+    using DataPoint         = typename Traits::DataPoint; /*!< DataPoint given by user via Traits                */     \
+    using Scalar            = typename DataPoint::Scalar; /*!< Scalar given by user via DataPoint                */     \
+    using VectorType        = typename DataPoint::VectorType; /*!< VectorType given by user via DataPoint            */ \
+    using IndexType         = typename Traits::IndexType; /*!< Type used to index points into the PointContainer */     \
+    using PointContainer    = typename Traits::PointContainer; /*!< Container for DataPoint used inside the KdTree   */ \
+    using IndexContainer    = typename Traits::IndexContainer; /*!< Container for indices used inside the KdTree     */ \
+    using IndexContainerRef = typename Traits::IndexContainerRef; /*!< Ref type to index container */
         WRITE_TRAITS
         using Base = NeighborGraphBufferBase<_Traits>;
 
@@ -145,7 +68,7 @@ namespace Ponca
      * \tparam _Traits Traits type providing the types and constants used by the KnnGraph. Must have the
      * same interface as the default traits type.
      *
-     * \see KnnGraphDefaultTraits for the trait interface documentation.
+     * \see NeighborGraphDefaultTraits for the trait interface documentation.
      *
      */
     template <typename _Traits>
