+++ cmake/PoncaConfigureSpatialPartitioning.cmake
@@ -13,8 +13,9 @@ set(ponca_SpatialPartitioning_INCLUDE
     "${PONCA_src_ROOT}/Ponca/src/SpatialPartitioning/KdTree/Iterator/kdTreeKNearestIterator.h"
     "${PONCA_src_ROOT}/Ponca/src/SpatialPartitioning/KdTree/Iterator/kdTreeNearestIterator.h"
     "${PONCA_src_ROOT}/Ponca/src/SpatialPartitioning/KdTree/Iterator/kdTreeRangeIterator.h"
+    "${PONCA_src_ROOT}/Ponca/src/SpatialPartitioning/NeighborGraph/neighborGraph.h"
+    "${PONCA_src_ROOT}/Ponca/src/SpatialPartitioning/NeighborGraph/neighborGraphTraits.h"
     "${PONCA_src_ROOT}/Ponca/src/SpatialPartitioning/KnnGraph/knnGraph.h"
-    "${PONCA_src_ROOT}/Ponca/src/SpatialPartitioning/KnnGraph/knnGraphTraits.h"
     "${PONCA_src_ROOT}/Ponca/src/SpatialPartitioning/KnnGraph/Query/knnGraphKNearestQuery.h"
     "${PONCA_src_ROOT}/Ponca/src/SpatialPartitioning/KnnGraph/Query/knnGraphRangeQuery.h"
     "${PONCA_src_ROOT}/Ponca/src/SpatialPartitioning/KnnGraph/Iterator/knnGraphRangeIterator.h"
