+++ tests/src/queries_knearest.cpp
@@ -125,7 +125,7 @@ void buildAndTestKnnGraph(KdTree& kdtree, const int k, const std::string& name =
 #endif
 
     // Test the KnnGraph with raw memory pointers
-    using KnnGraphPointerStatic = StaticKnnGraphBase<KnnGraphPointerTraits<P>>;
+    using KnnGraphPointerStatic = StaticKnnGraphBase<NeighborGraphPointerTraits<P>>;
     auto knngraphBuffers        = knnGraph.buffers(); // Buffer that use STL-like containers
     // Convert previous KnnGraph to pointers
     P* pts = new P[points.size()];
